requirements.txt
Flask
flask-restx
autopep8
python-dotenv
konfik
flask_sqlalchemy
pymysql
cryptography
flask_mongoengine
flask_marshmallow
marshmallow-sqlalchemy
Marshmallow-Mongoengine

connexion[swagger-ui]

# testing
pytest
deepdiff

# JWT
python-jose[cryptography]

# Kafka
kafka-python